requirements.txt
pandas
scikit-learn
numpy
fastapi
uvicorn
requests
python-dotenv
httpx
python-multipart
